Fix ImGui assert
The ImGui documentation states that ImGuiCond should be treated as a regular enum and cannot be combined using binary operators, and when running with ImGui assertions enabled, this line asserts.

I think Appearing is probably the desired value here, since "FirstUseEver" is "Set the variable if the object/window has no persistently saved data (no entry in .ini file)" and "Appearing" is "Set the variable if the object/window is appearing after being hidden/inactive (or the first time)".

Since from ImGui's point of view the window never actually hides (ImGui.Begin is never called with p_open=false, since we short circuit out of the Draw before then if the window is hidden), this has the effect of only setting the size of the window on the first frame that the window is visible, either after Dalamud first loads or the plugin (re)loads, I'm not entirely sure which.